+++ src/Service/AppSync/src/Result/ListResolversResponse.php
@@ -2,6 +2,7 @@
 
 namespace AsyncAws\AppSync\Result;
 
+use AsyncAws\AppSync\ValueObject\AppSyncRuntime;
 use AsyncAws\AppSync\ValueObject\CachingConfig;
 use AsyncAws\AppSync\ValueObject\LambdaConflictHandlerConfig;
 use AsyncAws\AppSync\ValueObject\PipelineConfig;
@@ -45,6 +46,14 @@ class ListResolversResponse extends Result
 
         $this->resolvers = empty($data['resolvers']) ? [] : $this->populateResultResolvers($data['resolvers']);
         $this->nextToken = isset($data['nextToken']) ? (string) $data['nextToken'] : null;
+    }
+
+    private function populateResultAppSyncRuntime(array $json): AppSyncRuntime
+    {
+        return new AppSyncRuntime([
+            'name' => (string) $json['name'],
+            'runtimeVersion' => (string) $json['runtimeVersion'],
+        ]);
     }
 
     private function populateResultCachingConfig(array $json): CachingConfig
@@ -115,6 +124,8 @@ class ListResolversResponse extends Result
             'syncConfig' => empty($json['syncConfig']) ? null : $this->populateResultSyncConfig($json['syncConfig']),
             'cachingConfig' => empty($json['cachingConfig']) ? null : $this->populateResultCachingConfig($json['cachingConfig']),
             'maxBatchSize' => isset($json['maxBatchSize']) ? (int) $json['maxBatchSize'] : null,
+            'runtime' => empty($json['runtime']) ? null : $this->populateResultAppSyncRuntime($json['runtime']),
+            'code' => isset($json['code']) ? (string) $json['code'] : null,
         ]);
     }