+++ src/Service/Ecr/src/EcrClient.php
@@ -68,6 +68,7 @@ class EcrClient extends AbstractApi
             case 'eu-central-2':
             case 'eu-north-1':
             case 'eu-south-1':
+            case 'eu-south-2':
             case 'eu-west-1':
             case 'eu-west-2':
             case 'eu-west-3':