+++ src/Service/Kms/src/KmsClient.php
@@ -329,6 +329,7 @@ class KmsClient extends AbstractApi
             case 'eu-central-2':
             case 'eu-north-1':
             case 'eu-south-1':
+            case 'eu-south-2':
             case 'eu-west-1':
             case 'eu-west-2':
             case 'eu-west-3':