+++ src/Service/Lambda/src/LambdaClient.php
@@ -381,6 +381,7 @@ class LambdaClient extends AbstractApi
             case 'eu-central-2':
             case 'eu-north-1':
             case 'eu-south-1':
+            case 'eu-south-2':
             case 'eu-west-1':
             case 'eu-west-2':
             case 'eu-west-3':