+++ src/Service/S3/src/S3Client.php
@@ -1017,6 +1017,7 @@ class S3Client extends AbstractApi
             case 'eu-central-2':
             case 'eu-north-1':
             case 'eu-south-1':
+            case 'eu-south-2':
             case 'eu-west-1':
             case 'eu-west-2':
             case 'eu-west-3':